+++ NEWS.md
@@ -1,3 +1,20 @@
+# MODIS 1.2.8.9001 (2022-09-20)
+
+#### ✨ features and improvements
+
+  * Aligns `runGdal()` 'maskValue' with `gdalwarp` '-srcnodata' option, thereby 
+    allowing to specify band-specific masking values or ignore intrinsic no-data
+    settings altogether (#70)
+
+#### 🐛 bug fixes
+
+#### 💬 documentation etc
+
+#### 🍬 miscellaneous
+
+  * Removes obsolete Python code initially used for value masking
+
+
 # MODIS 1.2.8 (2022-09-07)
 
 #### ✨ features and improvements